Restructure function to avoid circular imports
